Refactor battlefield
